Updated the Fricking Cyberdemon
It's now a lot tougher to deal with (and hopefully a lot more fun), but also has less health.
Added new attacks and decreased the time it can be stunned in its pain state.
It also has a chance of dropping energy pickups if hit with any normal attack.
Updated the arena.