Caliburn tweaks
Sonic can now cancel Boost into Sword Swing, and it'll deal MUCH more damage if he does so.
To counter this, and to make it work properly, Sonic's Boost will only output a knockback hitscan attack (or damaging hitscan attack if transformed) on the first frame; continuing to hold the Boost will not output any damage.