Updated stuff
New death message for when your Style Rank is A or higher.
Style Rank and Style Score have been moved closer to the top of the screen, to make room for ZDoom's Print function.
Map lumps have been changed to be more similar to one another. This is more a developer-side change.