Updated level structuring
In Station Square, there will now be a teleporter for each of the three levels (Sunset Gardens, Marble Madness, Oceanic Fortress), but you'll need to complete the previous levels to access either of the latter two.
Upon completing a level, you'll be taken back to Station Square, and you'll be able to access the next level's teleporter.
. . .of course, you can just `map` your way there.