Homing Assault tweaks
The Homing Assault now uses more or less the same code as the updated Homing Attack.
However, it will now cause an AoE attack when used. It's also just quite a bit stronger
In addition, a menu option has been added to switch the Homing Assault button to Homing Assault, just so it's less confusing for people used to Sonic's Homing Attack input.