Reworked Sword Dash
The input is now Forward just before Fire (aka. Forward just before Sword Slash), similar to the input for Accel Boost, instead of just pressing Fire mid-air.
This means it's a lot harder to perform accidentally, and it now has a more advanced input in a similar vein to Quick Dash.
To balance that out, the dash is a fair bit stronger, and it no longer requires the player to be airborne.
I'll put an entry for it on the website in a little bit, which I really should have done to start with.