Updated score display
If the player is in Super Form and their score meets the target score, it will be set to Yellow instead of Gold, since Gold is the default score display colour if the player's in Super Form.
This also affects Dark Form due to sharing the same powerup, but it's not much of a huge deal.