Fix: Spin Dash flag issue
The CVAR that controls friction if Classic Sonic is Spin Dashing not being reset if the Spin Dash is not reset when another character is switched to is now fixed.
Whenever IsSpinDashing is true, but the player is not Classic Sonic (meaning the character can't POSSIBLY Spin Dash), then the CVAR is reset to false.
This still doesn't fix the issue if the Spin Dash flag reset is cancelled by pausing the game and you start a new game as Classic Sonic, though this fixes the bulk of the issue, namely the other characters not being able to reset the Spin Dash state. Since the friction will be fixed as Classic Sonic upon just doing a Spin Dash, no fix has been applied to Classic himself just yet.