Update: Title screen
Fixed the "Nintendo, HIRE THIS MAN!" easter egg. Turns out all that was needed was changing it from an OPEN script to an ENTER script. Because it turns out OPEN scripts don't take CVARs into account. I'm very smart(!)
Also, it's now easier to trigger - 1 in 10 instead of 1 in 20. Though to be fair it's easier to trigger regardless because you couldn't trigger it before.
Man, that would have been funny to see it happen to people at SAGE/SFGHQ, but hey, v1.4.1 players will see it now >:)